+++ app/forecasts/views.py
@@ -1,7 +1,9 @@
 """Viewset for displaying forecasts, and fetching point forecast data is lat,lon are specified"""
+from json import JSONDecodeError
 import logging
+
 from rest_framework import viewsets
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, APIException
 from rest_framework.response import Response
 
 from forecasts.forecasts import forecast_list
@@ -43,7 +45,16 @@ class ForecastViewSet(viewsets.ViewSet):
             data["longitude"] = "`lon` parameter not specified"
 
         if "lat" in request.query_params and "lon" in request.query_params:
-            time_series = forecast.point_forecast(lat, lon)
+            try:
+                time_series = forecast.point_forecast(lat, lon)
+            except JSONDecodeError as error:
+                logger.error(
+                    f"Error retrieving dataset due to a JSON decode error: {error}",
+                    exc_info=True, 
+                )
+                raise APIException(detail=f"Error retrieving dataset for forecast slug: {pk}")
+
+
 
             data["time_series"] = [
                 {"time": reading[0], "reading": reading[1]} for reading in time_series